Moved landing page into its own componenet.
